fix types, update docs, format, merge useLocalNav to useLayout too, fix typo in leftAside,
