+++ src/shared/shared.ts
@@ -8,6 +8,7 @@ export type {
   LocaleConfig,
   LocaleSpecificConfig,
   MarkdownEnv,
+  MenuItem,
   PageData,
   PageDataPayload,
   SiteData,
