+++ theme.d.ts
@@ -1,8 +1,9 @@
 // so that users can do `import DefaultTheme from 'vitepress/theme'`
 
-import type { DefineComponent } from 'vue'
+import type { ComputedRef, DefineComponent, ShallowRef } from 'vue'
 import type { EnhanceAppContext } from './dist/client/index.js'
 import type { DefaultTheme } from './types/default-theme.js'
+import type { MenuItem } from './types/shared.js'
 
 export type { DefaultTheme } from './types/default-theme.js'
 
@@ -12,8 +13,29 @@ declare const theme: {
 }
 
 export default theme
-export declare const useSidebar: () => DefaultTheme.DocSidebar
-export declare const useLocalNav: () => DefaultTheme.DocLocalNav
+
+export declare const useLayout: () => {
+  isHome: ComputedRef<boolean>
+
+  sidebar: ComputedRef<DefaultTheme.SidebarItem[]>
+  sidebarGroups: ComputedRef<DefaultTheme.SidebarItem[]>
+  hasSidebar: ComputedRef<boolean>
+  isSidebarEnabled: ComputedRef<boolean>
+
+  hasAside: ComputedRef<boolean>
+  leftAside: ComputedRef<boolean>
+
+  /**
+   * The outline headers of the current page.
+   */
+  headers: ShallowRef<MenuItem>
+  /**
+   * Whether the current page has a local nav. Local nav is shown when the
+   * "outline" is present in the page. However, note that the actual
+   * local nav visibility depends on the screen width as well.
+   */
+  hasLocalNav: ComputedRef<boolean>
+}
 
 // TODO: add props for these
 export declare const VPBadge: DefineComponent
