+++ types/shared.d.ts
@@ -205,3 +205,8 @@ export interface MarkdownEnv {
   realPath?: string
   localeIndex?: string
 }
+
+export type MenuItem = Omit<Header, 'slug' | 'children'> & {
+  element: HTMLHeadElement
+  children?: MenuItem[]
+}
